Move BackgroundModel to render-app and rename packages
